Move `update_dendrite_boost_stregth` to `boosted_dendrites.py`.
