+++ nupic/research/frameworks/vernon/mixins/update_dendrite_boost_strength.py
@@ -20,24 +20,14 @@
 # ----------------------------------------------------------------------
 
 
-from nupic.research.frameworks.dendrites import BoostedDendritesBase
+from nupic.research.frameworks.dendrites import (
+    BoostedDendritesBase,
+    update_dendrite_boost_stregth,
+)
 
 __all__ = [
     "UpdateDendriteBoostStrength"
 ]
-
-
-def update_dendrite_boost_stregth(m):
-    """Function used to update BoostedDendritesBase modules boost strength. This is
-    typically done during training at the beginning of each epoch.
-
-    Call using :meth:`torch.nn.Module.apply` after each epoch if required
-    For example: ``m.apply(update_boost_strength)``
-
-    :param m: BoostedDendritesBase module
-    """
-    if isinstance(m, BoostedDendritesBase):
-        m.update_boost_strength()
 
 
 class UpdateDendriteBoostStrength:
